Events

Events

Events are computations $\,α,β,γ,...\,$ at times $\,t_1,t_2,t_3,...$ Those computations are called actions.

Actions

Actions are Julia functions or expressions to be executed later:

DiscreteEvents.Action — Constant.

Action

An action is either a Function or an Expr or a Tuple of them. It can be scheduled in an event for later execution.

source

DiscreteEvents.fun — Function.

fun(f::Function, args...; kwargs...)

Save a function f and its arguments in a closure for later execution.

Arguments

The arguments args... and keyword arguments kwargs... to fun are passed to f at execution but may change their values between beeing captured in fun and fs later execution. If f needs their current values at execution time there are two possibilities:

fun can take funs, function closures, symbols or expressions at the place of values or variable arguments. They are evaluated at event time just before being passed to f. There is one exception: if f is an event!, its arguments are passed on unevaluated.

A mutable type argument (Array, struct ...) is always current. You can also change its content from within a function.

If using Symbols or Expr in fun you get a one time warning. They are evaluated at global scope in Module Main only and therefore cannot be used by other modules.

source

Evaluating expressions is slow!

The use of expressions (Expr) and symbols (Symbol) in actions should be avoided. You will get a one time warning if you use them. They can be replaced easily by funs or function closures.

They are evaluated at global scope in Module Main only. Other modules using DiscreteEvents cannot use Expr in events and have to use functions.

Actions can be combined into tuples:

julia> using DiscreteEvents

julia> a = 1
1

julia> println(a) isa Action        # a function call is not an action
1
false

julia> fun(println, a) isa Action   # wrapped in a fun it is an action
true

julia> isa(()->println(a), Action)  # an anonymous function
true

julia> :(println(a)) isa Action     # an expression
true

julia> (()->println(a), fun(println, a), :(println(a))) isa Action # a tuple of them
true

Actions can be scheduled for execution

at given clock times and

under specified conditions.

Timed events

Actions can be scheduled as events at given times:

DiscreteEvents.Timing — Type.

Timing

Enumeration type for scheduling events and timed conditions:

at: schedule an event at a given time,

after: schedule an event a given time after current time,

every: schedule an event every given time from now on,

before: a timed condition is true before a given time,

until: delay until t.

source

DiscreteEvents.event! — Method.

event!([clk], ex, t; <keyword arguments>)
event!([clk], ex, T, t; <keyword arguments>)

Schedule an event for a given simulation time.

Arguments

clk<:AbstractClock: clock, it not supplied, the event is scheduled to 𝐶,

ex<:Action: an expression or function or a tuple of them,

T::Timing: a timing, one of at, after or every,

t<:Number: simulation time, if t < clk.time set t = clk.time,

Keyword arguments

cycle<:Number=0.0: repeat cycle time for an event,

cid::Int=clk.id: if cid ≠ clk.id, assign the event to the parallel clock with id == cid. This overrides spawn,

spawn::Bool=false: if true, spawn the event at other available threads,

sync::Bool=false: if true, force a synchronization of all parallel clocks before executing the event.

Examples

julia> using DiscreteEvents, Unitful

julia> import Unitful: s, minute, hr

julia> myfunc(a, b) = a+b
myfunc (generic function with 1 method)

julia> event!(𝐶, fun(myfunc, 1, 2), 1) # a 1st event to 1

julia> event!(𝐶, fun(myfunc, 2, 3), 1) #  a 2nd event to the same time

julia> event!(𝐶, fun(myfunc, 3, 4), 1s)
Warning: clock has no time unit, ignoring units

julia> setUnit!(𝐶, s)
0.0 s

julia> event!(𝐶, fun(myfunc, 4, 5), 1minute)

julia> event!(fun(myfunc, 5, 6), after, 1hr)

source

Conditional events

Actions can be scheduled as events under given conditions:

DiscreteEvents.event! — Method.

event!([clk], ex, cond; <keyword arguments>)

Schedule ex as a conditional event, conditions cond get evaluated at each clock tick.

Arguments

clk<:AbstractClock: if no clock is supplied, the event is scheduled to 𝐶,

ex<:Action: an expression or function or a tuple of them,

cond<:Action: a condition is true if all functions or expressions therein return true,

cid::Int=clk.id: assign the event to the parallel clock cid. This overrides spawn,

spawn::Bool=false: if true, spawn the event at other available threads.

Examples

julia> using DiscreteEvents

julia> c = Clock()   # create a new clock
Clock 1: state=:idle, t=0.0, Δt=0.01, prc:0
  scheduled ev:0, cev:0, sampl:0

julia> event!(c, fun((x)->println(tau(x), ": now I'm triggered"), c), fun(>=, fun(tau, c), 5))

julia> run!(c, 10)   # sampling is not exact, so it takes 502 sample steps to fire the event
5.009999999999938: now I'm triggered
"run! finished with 0 clock events, 502 sample steps, simulation time: 10.0"

source

Use inequalities to express conditions

For conditions you should prefer inequalities like <, ≤, ≥, > to equality == in order to make sure that a condition can be detected, e.g. tau() ≥ 100 is preferable to tau() == 100.

Continuous sampling

Actions can be registered for sampling and are then executed "continuously" at each clock increment Δt. The default clock sample rate Δt is 0.01 time units.

DiscreteEvents.sample_time! — Function.

sample_time!([clk::Clock], Δt::N) where {N<:Number}

Set the clock's sample rate starting from now (tau(clk)).

Arguments

clk::Clock: if not supplied, set the sample rate on 𝐶,

Δt::N: sample rate, time interval for sampling

source

DiscreteEvents.periodic! — Function.

periodic!([clk], ex, Δt; spawn)

Register a function or expression for periodic execution at the clock`s sample rate.

Arguments

clk<:AbstractClock: if not supplied, it registers on 𝐶,

ex<:Action: an expression or function or a tuple of them,

Δt<:Number=clk.Δt: set the clock's sampling rate, if no Δt is given, it takes the current sampling rate, if that is 0, it calculates one,

spawn::Bool=false: if true, spawn the periodic event to other available threads.

source

Events and variables

Actions often depend on data or modify it. The data may change between the definition of an action and its later execution. If an action uses a mutable variable like an array or a mutable struct, it gets current data at event time and it is fast. If the action modifies the data, this is the best way to do it:

julia> using DiscreteEvents

julia> a = [1]                  # define a mutable variable a
1-element Array{Int64,1}:
 1

julia> f(x; y=2) = (x[1] += y)  # define a function f
f (generic function with 1 method)

julia> ff = fun(f, a);          # enclose f and a in a fun ff

julia> a[1] += 1                # modify a
2

julia> ff()                     # execute ff
4

julia> a[1]                     # a has been modified correctly
4

If for some reason and against better advise you want to work with global variables, there are several ways to deal with them:

julia> g(x; y=1) = x+y                 # define a function g
g (generic function with 1 method)

julia> x = 1;                          # define a global variable x = 1

julia> gg = fun(g, :x, y=2);           # pass x as a symbol to g

julia> x += 1                          # increment x
2

julia> 2
2

julia> gg()                            # now g gets a current x and gives a warning
┌ Warning: Evaluating expressions is slow, use functions instead
└ @ DiscreteEvents ~/build/pbayer/DiscreteEvents.jl/src/fclosure.jl:28
4

julia> hh = fun(g, fun(()->x), y=3);   # reference x with an anonymous fun

julia> x += 1                          # increment x
3

julia> hh()                            # g gets again a current x
6

julia> ii = fun(g, ()->x, y=4);        # reference x with an anonymous function

julia> x += 1                          # increment x
4

julia> ii()                            # g gets an updated x
8

If you want to modify a global variable, you have to use the global keyword inside your function.