Internals

Internals

Clocks

Simulate.jl contains two clock types: Clock and ActiveClock. Both are implemented as state machines. The implementation functions and types are not exported.

We have some definitions for them to work.

Simulate.AbstractClock — Type.

supertype for clocks in Simulate.jl

source

States

Defined clock states.

Simulate.SState — Type.

supertype for states

source

Simulate.Undefined — Type.

a state machine is undefined (after creation)

source

Simulate.Idle — Type.

a state machine is idle

source

Simulate.Empty — Type.

a state machine is empty

source

Simulate.Busy — Type.

a state machine is busy

source

Simulate.Halted — Type.

a state machine is halted

source

Events

Defined events.

Simulate.SEvent — Type.

supertype for events

source

Simulate.Init — Type.

Init(info): Init event with some info.

source

Simulate.Setup — Type.

Setup(vars::Array{Symbol,1}, scope::Module): setup a logger with some info.

source

Simulate.Switch — Type.

Switch(to): switch to some other mode

source

Simulate.Log — Type.

Log(): record command for logging

source

Simulate.Step — Type.

Step(): command

source

Simulate.Run — Type.

Run(): command

source

Simulate.Reset — Type.

Reset(): command

source

Simulate.Query — Type.

Query(): command, causes an active clock to send its clock data.

source

Simulate.Diag — Type.

Diag(): command, causes an active clock to send the last stacktrace.

source

Simulate.Response — Type.

Response(): response from an active clock

source

Simulate.Register — Type.

Register(): command

source

Simulate.Sync — Type.

Sync(): command

source

Simulate.Start — Type.

Start(): command

source

Simulate.Stop — Type.

Stop(): command

source

Simulate.Resume — Type.

Resume(): command

source

Simulate.Clear — Type.

Clear(): command

source

Multithreading

Internal functions used for multithreading.

Simulate.start_threads — Function.

start_threads(f::Function, mul::Int=3)

Start a task on each available thread (other than 1).

Arguments

f::Function: function to start, has to take a channel as its only argument,

mul::Int=3: startup multiplication factor,

source

Simulate.activeClock — Function.

activeClock(ch::Channel)

Operate an active clock on a given channel. This is its event loop.

source

Simulate.spawnid — Function.

spawnid(clk::Clock) :: Int

Return a random number out of the thread ids of all available parallel clocks. This is used for spawning tasks or events to them.

Note

This function may be used for workload balancing between threads in the future.

source

Transition functions

The internal clock transition function has several methods for different combinations of clock types, states and events:

Simulate.step! — Function.

step!(A::AbstractClock, q::SState, σ::SEvent)

Default transition for clock state machines.

This is called if no otherwise defined transition occurs.

Arguments

A::AbstractClock: state machine for which a transition is called

q::SState: state of the state machine

σ::SEvent: event, triggering the transition

source

step!(clk::Clock, ::Undefined, ::Init)

initialize a clock.

source

step!(clk::Clock, ::Undefined, σ::Union{Step,Run})

if uninitialized, initialize and then Step or Run.

source

step!(c::Clock, ::Union{Idle,Busy,Halted}, ::Step)

step forward to next tick or scheduled event.

At a tick evaluate 1) all sampling functions or expressions, 2) all conditional events, then 3) if an event is encountered, trigger the event.

The internal clock times c.tev and c.tn are always at least c.time.

source

step!(clk::Clock, ::Idle, σ::Run)

Run a simulation for a given duration.

The duration is given with Run(duration). Call scheduled events and evaluate sampling expressions at each tick in that timeframe.

source

step!(clk::Clock, ::Busy, ::Stop)

Stop the clock.

source

step!(clk::Clock, ::Halted, ::Resume)

Resume a halted clock.

source

step!(clk::Clock, q::SState, σ::SEvent)

catch all step!-function.

source

Run an active clock for a given duration.

source

Other internal types and functions

Simulate.SimEvent — Type.

SimEvent{T<:Union{SimFunction,Expr,Tuple}}

A simulation event is a SimFunction or an expression or a tuple of them to be executed at an event time.

Arguments, fields

ex::T: a SimFunction or an expression or a tuple of them,

scope::Module: evaluation scope,

t::Float64: event time,

Δt::Float64: repeat rate with for repeating events.

source

Simulate.SimCond — Type.

SimCond{S<:Union{SimFunction,Expr,Tuple}, T<:Union{SimFunction,Expr,Tuple}}

A condition to be evaluated repeatedly with expressions or functions to be executed if conditions are met.

Arguments, fields

cond::S: a SimFunction or an expression or a tuple of them,

ex::T: a SimFunction or an expression or a tuple of them,

scope::Module: evaluation scope

source

Simulate.Sample — Type.

Sample{T<:Union{SimFunction,Expr}}

A sampling function or expression is called at sampling time.

Arguments, fields

ex::SimExpr: expression or SimFunction to be called at sample time

scope::Module: evaluation scope

source

Simulate.sconvert — Function.

sconvert(ex::Union{SimExpr, Tuple, Vector})::Tuple{Vararg{SimExpr}}

convert a SimExpr or an array or a tuple of it to a Tuple{Vararg{SimExpr}}

source

Simulate.nextevent — Function.

nextevent(c::Clock)

Return the next scheduled event.

source

Simulate.nextevtime — Function.

nextevtime(c::Clock)

Return the internal time (unitless) of next scheduled event.

source

Simulate.checktime — Function.

checktime(clk::Clock, t::Number)::Float64

check t given according to clock settings and return a Float64 value

source

Simulate.evaluate — Function.

catchall function: forward the value y

source

recursive call to sfExec for a nested SimFunction.

source

evaluate a symbol or expression and give a warning.

source

Simulate.evExec — Function.

evExec(ex::Union{SimExpr, Tuple, m::Module=Main)

Forward an event's SimFunctions or expressions to further execution or evaluation.

Return

the evaluated value or a tuple of evaluated valuesch.

source

Simulate.sfExec — Function.

execute a SimFunction

source

Forward an expression to evaluate.

source

Simulate.setTimes — Function.

setTimes(clk::Clock)

set clock times for next event or sampling action. The internal clock times clk.tev and clk.tn must always be set to be at least clk.time.

source

Simulate.do_event! — Function.

do_event!(c::Clock)

Execute or evaluate the next timed event on a clock c.

source

Simulate.do_tick! — Function.

do_tick!(s::Schedule)

First execute all sampling expressions in a schedule, then evaluate all conditional events and if conditions are met, execute them.

source

Simulate.do_step! — Function.

do_step!(c::Clock)

step forward to next tick or scheduled event.

source

Simulate.do_run! — Function.

do_run(c::Clock, Δt::Float64)

Run a clock for a time Δt.

source

Simulate.startup! — Function.

startup!(p::SimProcess, cycles::Number, spawn::Bool)

Start a SimProcess as a task in a loop.

source

Simulate.register! — Function.

register!(clk::Clock, p::SimProcess)

Register a SimProcess to a clock. Check its id and change it apropriately.

source

Simulate.wakeup — Function.

wakeup a process waiting for a Condition

source

Simulate.init! — Function.

init!(clk::Clock)

initialize a clock.

source

Simulate.loop — Function.

loop(p::SimProcess, start::Channel, cycles::Number)

Put a SimProcess in a loop which can be broken by a SimException.

Arguments

p::SimProcess:

start::Channel: a channel to ensure that a process starts,

cycles=Inf: determine, how often the loop should be run.

source

Simulate.scale — Function.

scale(n::Number)::Float64

calculate the scale from a given number

source