User guide

User guide

Installation

Simulate.jl runs on Julia versions ≥ v"1.0". Multithreading requires Julia ≥ v"1.3".

Simulate.Simulate — Module.

Simulate

A Julia package for discrete event simulation.

The current stable, registered version is installed with

pkg> add Simulate

The development version is installed with:

pkg> add("https://github.com/pbayer/Simulate.jl")

source

Simulate.version — Constant.

version

Gives the package version:

julia> using Simulate

julia> Simulate.version
v"0.3.0"

source

Clocks

A clock in Simulate.jl is an active object residing in a thread and registers function calls or expressions as events, schedules them at a given time or under a given condition and executes them at their time or if conditions are met.

Clock and ActiveClocks have virtual (simulation) time. A Clock can control and synchronizes with ActiveClocks on other threads. It can be started and run for a given time.

RTClocks have real (system) time and operate independently from each other and other clocks. They run continuously at given clock ticks and execute scheduled events if their time becomes due.

Simulate.Clock — Type.

Clock(Δt::Number=0; t0::Number=0, unit::FreeUnits=NoUnits)

Create a new simulation clock.

Arguments

Δt::Number=0: time increment. If no Δt is given, the simulation doesn't tick, but jumps from event to event. Δt can be set later with sample_time!.

t0::Number=0: start time for simulation

unit::FreeUnits=NoUnits: clock time unit. Units can be set explicitely by setting e.g. unit=minute or implicitly by giving Δt as a time or else setting t0 to a time, e.g. t0=60s.

Fields

id::Int: thread on which the clock is running,

state::ClockState: clock state,

time::Float64: clock time,

unit::FreeUnits: time unit,

end_time::Float64: end time for simulation,

Δt::Float64: sampling time, timestep between ticks,

ac::Vector{ClockChannel}: active clocks running on parallel threads,

sc::Schedule: the clock schedule (events, cond events and sampling),

processes::Dict{Any, Prc}: registered Prces,

tn::Float64: next timestep,

tev::Float64: next event time,

evcount::Int: event counter,

scount::Int: sample counter

Examples

julia> using Simulate, Unitful

julia> import Unitful: s, minute, hr

julia> c = Clock()                 # create a unitless clock (standard)
Clock thread 1 (+ 0 ac): state=Simulate.Undefined(), t=0.0 , Δt=0.0 , prc:0
  scheduled ev:0, cev:0, sampl:0

julia> c = Clock(1s, unit=minute)  # create a clock with units, does conversions automatically
Clock thread 1 (+ 0 ac): state=Simulate.Undefined(), t=0.0 minute, Δt=0.01667 minute, prc:0
  scheduled ev:0, cev:0, sampl:0

julia> c = Clock(1s)               # create a clock with implicit unit setting
Clock thread 1 (+ 0 ac): state=Simulate.Undefined(), t=0.0 s, Δt=1.0 s, prc:0
  scheduled ev:0, cev:0, sampl:0

julia> c = Clock(t0=60s)           # another example of implicit unit setting
Clock thread 1 (+ 0 ac): state=Simulate.Undefined(), t=60.0 s, Δt=0.0 s, prc:0
  scheduled ev:0, cev:0, sampl:0

julia> c = Clock(1s, t0=1hr)       # if given times with different units, Δt takes precedence
Clock thread 1 (+ 0 ac): state=Simulate.Undefined(), t=3600.0 s, Δt=1.0 s, prc:0
  scheduled ev:0, cev:0, sampl:0

source

Simulate.ActiveClock — Type.

ActiveClock(clock::Clock, master::Ref{Clock},
            cmd::Channel{ClockEvent}, ans::Channel{ClockEvent})

A thread specific clock which can be operated via a channel.

Fields

clock::Clock: the thread specific clock,

master::Ref{Clock}: a pointer to the master clock (on thread 1),

cmd::Channel{ClockEvent}: the command channel from master,

ans::Channel{ClockEvent}: the response channel to master.

id::Int: the id in master's ac array,

thread::Int: the thread, the active clock runs on.

Note

You should not setup an ActiveClock explicitly. Rather this is done implicitly by fork!ing a Clock to other available threads or directly with PClock. It then can be accessed via pclock as in the following example.

Example

julia> using Simulate

julia> clk = Clock()
Clock thread 1 (+ 0 ac): state=Simulate.Undefined(), t=0.0 , Δt=0.0 , prc:0
  scheduled ev:0, cev:0, sampl:0

julia> fork!(clk)

julia> clk    #  ⬇ you got 3 parallel active clocks
Clock thread 1 (+ 3 ac): state=Simulate.Undefined(), t=0.0 , Δt=0.0 , prc:0
  scheduled ev:0, cev:0, sampl:0

julia> clk = PClock()
Clock thread 1 (+ 3 ac): state=Simulate.Undefined(), t=0.0 , Δt=0.01 , prc:0
  scheduled ev:0, cev:0, sampl:0

julia> pclock(clk, 1)    # get access to the 1st active clock (on thread 2)
Active clock 1 on thrd 2: state=Simulate.Idle(), t=0.0 , Δt=0.01 , prc:0
   scheduled ev:0, cev:0, sampl:0

source

Simulate.RTClock — Type.

RTClock{T <: ClockEvent} <: AbstractClock

Real time clocks use system time for time keeping and are controlled over a channel. They are independent from each other and other clocks and run asynchronously as tasks on arbitrary threads. Multiple real time clocks can be setup with arbitrary frequencies to do different things.

Fields

clock::Clock:

cmd::Channel{T}:

back::Channel{T}:

id::Int:

thread::Int:

t0::Float64:

source

Clocks have the following substructures:

Simulate.Schedule — Type.

Schedule()

A Schedule contains events, conditional events and sampling functions to be executed or evaluated on the clock's time line.

Fields

events::PriorityQueue{DiscreteEvent,Float64}: scheduled events,

cevents::Array{DiscreteCond,1}: conditional events to evaluate at each tick,

samples::Array{Sample,1}: sampling expressions to evaluate at each tick,

source

Simulate.ClockChannel — Type.

ClockChannel

Provide a channel to an active clock or a real time clock.

Fields

ref::Ref{Task}: a pointer to an active clock,

ch::Channel: a communication channel to an active clock,

id::Int: the thread id of the active clock.

done::Bool: flag indicating if the active clock has completed its cycle.

source

You can set time units and query the current clock time. There is a default clock 𝐶, which can be used for experimental work. RTC can be used to setup and control real time Clocks.

Simulate.setUnit! — Function.

setUnit!(clk::Clock, new::FreeUnits)

set a clock to a new time unit in Unitful. If necessary convert current clock times to the new unit.

Arguments

clk::Clock

new::FreeUnits: new is one of ms, s, minute or hr or another Unitful Time unit.

Examples

julia> using Simulate, Unitful

julia> import Unitful: Time, s, minute, hr

julia> c = Clock(t0=60)     # setup a new clock with t0=60
Clock thread 1 (+ 0 ac): state=Simulate.Undefined(), t=60.0 , Δt=0.0 , prc:0
  scheduled ev:0, cev:0, sampl:0

julia> tau(c) # current time is 60.0 NoUnits
60.0

julia> setUnit!(c, s)       # set clock unit to Unitful.s
60.0 s

julia> tau(c) # current time is now 60.0 s
60.0 s

julia> setUnit!(c, minute)  # set clock unit to Unitful.minute
1.0 minute

julia> tau(c)               # current time is now 1.0 minute
1.0 minute

julia> isa(tau(c), Time)
true

julia> uconvert(s, tau(c))  # ... which can be converted to other time units
60.0 s

julia> tau(c).val           # it has a value of 1.0
1.0

julia> c.time               # internal clock time is set to 1.0 (a Float64)
1.0

julia> c.unit               # internal clock unit is set to Unitful.minute
minute

source

Simulate.tau — Function.

tau(clk::Clock=𝐶)

Return the current simulation time.

Examples

julia> using Simulate

julia> reset!(𝐶)
"clock reset to t₀=0.0, sampling rate Δt=0.0."
julia> tau() # gives the central time
0.0

source

Simulate.𝐶 — Constant.

𝐶

𝐶 (𝐶 = \itC+tab) is the default simulation clock. If you do one simulation at a time, you can use it for time keeping.

Examples

julia> using Simulate

julia> reset!(𝐶)
"clock reset to t₀=0.0, sampling rate Δt=0.0."

julia> 𝐶  # default clock
Clock thread 1 (+ 0 ac): state=Simulate.Idle(), t=0.0 , Δt=0.0 , prc:0
  scheduled ev:0, cev:0, sampl:0

source

Simulate.RTC — Constant.

RTC::Vector{ClockChannel}

Real time clocks are registered to RTC and can be accessed and operated through it.

source

You can create a clock with parallel active clocks on all available threads or fork existing clocks to other threads or collapse them if no longer needed. You can get direct access to parallel clocks and diagnose them.

Simulate.PClock — Function.

PClock(Δt::Number=0.01; t0::Number=0, unit::FreeUnits=NoUnits)

Setup a clock with parallel clocks on all available threads.

Arguments

Δt::Number=0.01: time increment. For parallel clocks Δt has to be > 0. If given Δt ≤ 0 it is set to 0.01.

t0::Number=0: start time for simulation

unit::FreeUnits=NoUnits: clock time unit. Units can be set explicitely by setting e.g. unit=minute or implicitly by giving Δt as a time or else setting t0 to a time, e.g. t0=60s.

Note

Processes on multiple threads are possible in Julia ≥ 1.3 and with JULIA_NUM_THREADS > 1.

source

Simulate.fork! — Function.

fork!(master::Clock)

Establish copies of a clock on all parallel threads and operate them as active clocks under control of the master clock.

Arguments

master::Clock: the master clock, must be on thread 1

source

Simulate.pclock — Function.

pclock(clk::Clock, id::Int ) :: AbstractClock
pclock(ac::ActiveClock, id::Int ) :: AbstractClock

Get a parallel clock to a given clock.

Arguments

master::Clock: a master clock or

ac::ActiveClock: an active clock,

id::Int=threadid(): thread id, defaults to the caller's current thread.

Returns

the master Clock if id==0,

a parallel ActiveClock else

source

Simulate.collapse! — Function.

collapse!(master::Clock)

Transfer the schedules of the parallel clocks to master and them stop them.

Note

If there are processes on other threads registered to parallel clocks, make sure that they aren't needed anymore before calling collapse. They are not transferred to and cannot be controlled by master.

source

Simulate.diag — Function.

diag(clk::Clock, id::Int)

Return the stacktrace from a parallel clock.

Arguments

clk::Clock: a master clock,

id::Int: the id of a parallel clock.

source

Note

Directly accessing the clock substructure of parallel ActiveClocks is possible but not recommended since it breaks parallel operation. The right way is to pass event!s to the ActiveClock-variable. The communication then happens over the channel to the ActiveClock as it should be.

Events

Julia expressions and functions can be scheduled as events on the clock's timeline and are triggered at a given simulation time or under conditions which may become true during simulation.

Functions and expressions can be scheduled for execution

at given clock times and

under specified conditions.

Simulate.AbstractEvent — Type.

supertype for events

source

Simulate.DiscreteEvent — Type.

DiscreteEvent{T<:Action} <: AbstractEvent

A discrete event is a function or an expression or a tuple of them to be executed at an event time.

Arguments, fields

ex::T: a function or an expression or a tuple of them,

scope::Module: evaluation scope,

t::Float64: event time,

Δt::Float64: repeat rate with for repeating events.

source

Simulate.DiscreteCond — Type.

DiscreteCond{S<:Action, T<:Action} <: AbstractEvent

A condition to be evaluated repeatedly with expressions or functions to be executed if conditions are met.

Arguments, fields

cond::S: a conditional function or an expression or a tuple of them (conditions must evaluate to Bool),

ex::T: a function or an expression or a tuple of them to be executed if conditions are met,

scope::Module: evaluation scope

source

Simulate.Timing — Type.

Timing

Enumeration type for scheduling events and timed conditions:

at: schedule an event at a given time,

after: schedule an event a given time after current time,

every: schedule an event every given time from now on,

before: a timed condition is true before a given time,

until: delay until t.

source

Simulate.fun — Function.

fun(f::Function, args..., kwargs...)

Saves a function and its arguments for later execution.

Arguments

fun can take any arguments. If you want f at execution time to have current arguments you can give symbols, expressions or other funs. They are then evaluated just before being passed to f. There is one exception: if f is an event!, its arguments are not evaluated before execution.

Evaluating symbols and expressions is slow

It should be avoided in time critical parts of applications. You will get a one time warning if you use that feature. See the Performance section in the documentation.

Returns

It returns a closure of f(args..., kwargs...) which has to be called with a Module argument (for evaluation scope of symbols and expressions).

Examples

source

Simulate.Action — Constant.

Action

An action is either an Expr or a Function or a Tuple of them.

source

Simulate.event! — Function.

event!

An event! is a Julia function or expression or a tuple of them to execute at a given event time or under given conditions.

source

Functions and expressions can be given to events on their own or in tuples, even mixed:

function events()
    event!(:(i += 1), after, 10)  # one expression
    event!(fun(f, 1, 2, 3, diff=pi), every, 1)  # one fun
    event!((:(i += 1), fun(g, j)), [:(tau() ≥ 50), fun(isready, input), :(a ≤ 10)]) # two funs under three conditions
end

All given functions or expressions are then called or evaluated at a given simulation time or when during simulation the given conditions become true.

Warning

Evaluating expressions or symbols at global scope is much slower than using functions and gives a one time warning. See Performance. This functionality may be removed entirely in a future version. (Please write an issue if you want to keep it.)

Note

Since conditions often are not met exactly you should prefer inequalities like <, ≤, ≥, > to equality == in order to get sure that a fulfilled condition can be detected, e.g. :(tau() ≥ 100) is preferable to :(tau() == 100).

Continuous sampling

Functions or expressions can register for sampling and are then executed "continuously" at each clock increment Δt.

Simulate.Sample — Type.

Sample{T<:Action} <: AbstractEvent

Sampling functions or expressions are called at sampling time.

Arguments, fields

ex<:Action: expression or function or a tuple of them to be called at sample time,

scope::Module: evaluation scope.

source

Simulate.sample_time! — Function.

sample_time!([clk::Clock], Δt::Number)

set the clock's sample rate starting from now (tau(clk)).

Arguments

clk::Clock: if not supplied, set the sample rate on 𝐶,

Δt::Number: sample rate, time interval for sampling

source

Simulate.periodic! — Function.

periodic!([clk::Clock], ex::T, Δt::Number=clk.Δt;
        scope::Module=Main, spawn=false) where {T<:Action}
periodic!(ac::ActiveClock, ex::T, Δt::Number=ac.clock.Δt; kw...) where {T<:Action}

Register a function or expression for periodic execution at the clock`s sample rate.

Arguments

clk::Clock, ac::ActiveClock: if not supplied, it samples on 𝐶,

ex<:Action: an expression or function or a tuple of them,

Δt::Number=clk.Δt: set the clock's sampling rate, if no Δt is given, it takes the current sampling rate, if that is 0, it calculates one,

scope::Module=Main: optional, an evaluation scope for a given expression.

source

Processes

Julia functions can be registered and run as processes. They follow another (the process-oriented) scheme and can be suspended and reactivated by the scheduler if they wait for something or delay. They must not (but are free to) handle and create events explicitly.

Simulate.Prc — Type.

Prc( id, f::Function, arg...; kw...)

Prepare a function to run as a process in a simulation.

Arguments, fields

id: some unique identification for registration,

task::Union{Task,Nothing}: a task structure,

clk::Union{AbstractClock,Nothing}: clock where the process is registered,

state::ClockState: process state,

f::Function: a function f(clk, arg...; kw...), must take clk as its first argument,

arg...: further arguments to f

kw...: keyword arguments to f

Note

A function started as a Prc most often runs in a loop. It has to give back control by e.g. doing a take!(input) or by calling delay! etc., which will yield it. Otherwise it will starve everything else!

Warn

That f nust take clk as first argument is a breaking change in v0.3!

source

Simulate.ClockException — Type.

ClockException(ev::ClockEvent, value=nothing)

Define a ClockException, which can be thrown to processes.

Arguments, fields

ev::ClockEvent: delivers an event to the interrupted task

value=nothing: deliver some other value

source

Simulate.process! — Function.

process!([clk::Clock], p::Prc, cycles=Inf; spawn::Bool=false)

Register a Prc to a clock, start it as an asynchronous process and return the id it was registered with. It can then be found under clk.processes[id].

Arguments

c::AbstractClock: Clock or ActiveClock, if not provided, the process runs under 𝐶,

p::Prc: it contains a function and its arguments,

cycles::Number=Inf: number of cycles the process should run,

spawn::Bool=false: if true, the process may be scheduled on another thread in parallel and registered to the thread specific clock.

Note

spawning a process is possible only with parallel clocks setup with PClock or fork!.

source

Simulate.interrupt! — Function.

interrupt!(p::Prc, ev::ClockEvent, value=nothing)

Interrupt a Prc by throwing a ClockException to it.

source

Simulate.stop! — Method.

Stop a Prc

source

Delay and wait …

Processes do not need to handle their events explicitly, but can call delay! or wait! or take! and put! … on their channels. This usually comes in handy. They are then suspended until certain conditions are met or requested resources are available.

Simulate.delay! — Function.

delay!(clk::Clock, t::Number)

Delay a process for a time interval t on the clock clk. Suspend the calling process until being reactivated by the clock at the appropriate time.

Arguments

clk::Clock: if not provided, the delay goes to 𝐶.

t::Number: the time interval for the delay.

source

delay!(clk::Clock, T::Timing, t::Number)

Used for delaying a process until a given time t.

Arguments

clk::Clock: if no clock is given, the delay goes to 𝐶,

T::Timing: only until is accepted,

t::Number: delay until time t if t > clk.time, else give a warning.

source

Simulate.wait! — Function.

wait!(clk::Clock, cond::Action; scope::Module=Main)

Wait on a clock for a condition to become true. Suspend the calling process until the given condition is true.

Arguments

clk::Clock: if no clock is supplied, the delay goes to 𝐶,

cond::Action: a condition is an expression or function or an array or tuple of them. It is true only if all expressions or functions therein return true,

scope::Module=Main: evaluation scope for given expressions.

source

Now

Processes in a simulation want their IO-operations to finish before the clock proceeds. Therefore they must enclose those operations in a now! call.

Simulate.now! — Function.

now!(clk::Clock, op::Action)

Tell the clock to execute an operation. Thus it cannot proceed before the op is finished.

Note

This is needed for IO-operations of tasks. IO-operations yield the task to the scheduler and the scheduler may invoke the clock before giving control back to the task. In that case the clock will proceed and the task has gone out of sync with the clock. Use now! to avoid this situation!

Arguments

clk::Clock,

op::Action: operation to execute.

source

Running simulations

If we run the clock, events are triggered, conditions are evaluated, sampling is done and delays are executed … We can also step through a simulation, stop, resume or reset it.

Simulate.reset! — Function.

reset!(clk::Clock, Δt::Number=0; t0::Number=0, hard::Bool=true, unit=NoUnits)

reset a clock

Arguments

clk::Clock

Δt::Number=0: time increment

t0::Float64=0 or t0::Time: start time

hard::Bool=true: time is reset, all scheduled events and sampling are deleted. If hard=false, then only time is reset, event and sampling times are adjusted accordingly.

unit=NoUnits: the Time unit for the clock after reset. If a Δt::Time is given, its Time unit goes into the clock Time unit. If only t0::Time is given, its Time unit goes into the clock time unit.

Examples

julia> using Simulate, Unitful

julia> import Unitful: s

julia> c = Clock(1s, t0=60s)
Clock thread 1 (+ 0 ac): state=Simulate.Undefined(), t=60.0 s, Δt=1.0 s, prc:0
  scheduled ev:0, cev:0, sampl:0

julia> reset!(c)
"clock reset to t₀=0.0, sampling rate Δt=0.0."

julia> c
Clock thread 1 (+ 0 ac): state=Simulate.Idle(), t=0.0 , Δt=0.0 , prc:0
  scheduled ev:0, cev:0, sampl:0

source

Simulate.incr! — Function.

incr!(clk::Clock)

Take one simulation step, execute the next tick or event.

source

Simulate.run! — Function.

run!(clk::Clock, duration::Number)

Run a simulation for a given duration. Call scheduled events and evaluate sampling expressions at each tick in that timeframe.

source

Simulate.stop! — Method.

stop!(clk::Clock)

Stop a running simulation.

source

Simulate.resume! — Function.

resume!(clk::Clock)

Resume a halted simulation.

source

Simulate.sync! — Function.

sync!(clk::Clock, to::Clock=𝐶)

Force a synchronization of two clocks. Change all registered times of clk accordingly. Convert or force clk.unit to to.unit.

source