Processes

Processes

Processes are typical event sequences implemented in a function. They are run by asynchronous tasks in a loop and get registered to a clock.

DiscreteEvents.Prc — Type.

Prc(id, f, arg...; kw...)

Prepare a function to run as a process registered to a clock.

Arguments, fields

id: some unique identification for registration,

f::Function: a function f(clk, arg...; kw...), must take clk (a Clock) as its first argument,

arg...: further arguments to f

kw...: keyword arguments to f

Fields

task::Union{Task,Nothing}: a task structure,

clk::Union{AbstractClock,Nothing}: clock where the process is registered,

state::ClockState: process state,

source

DiscreteEvents.process! — Function.

process!([clk], prc, cycles; spawn)

Register a Prc to a clock, start an asynchronous task executing the process function in a loop and return the id it was registered with. It can then be found under clk.processes[id].

Arguments

c<:AbstractClock: if not provided, the process runs under 𝐶,

prc::Prc: it contains a function and its arguments,

cycles<:Number=Inf: number of loop cycles the process should run,

spawn::Bool=false: if true, the process may be scheduled on another thread in parallel and registered to the thread specific clock.

Note

spawning a process is possible only with parallel clocks setup with PClock or fork!.

source

DiscreteEvents.interrupt! — Function.

interrupt!(p::Prc, ev::ClockEvent, value=nothing)

Interrupt a Prc by throwing a ClockException to it.

source

Processes must yield!

A process has to yield to Julia by e.g. doing a take!(input) or by calling delay! or wait!. Otherwise it will starve everything else!

Delay and wait …

In order to implement a process (an event sequence) functions can call delay! or wait! on the clock or take! and put! on channels. They are then suspended until a given time or until certain conditions are met or requested resources are available.

DiscreteEvents.delay! — Function.

delay!(clk, Δt)
delay!(clk, T, t)

Delay (suspend) a process for a time interval Δt on the clock clk.

Arguments

Δt<:Number: time interval,

T::Timing: only until is accepted,

t<:Number: delay until time t if t > clk.time, else give a warning.

source

DiscreteEvents.wait! — Function.

wait!(clk, cond)

Delay (suspend) a process on a clock clk until a condition has become true.

Arguments

cond<:Action: a condition is is true if all expressions or functions therein return true.

source

Processes use blocking calls and are not responsive!

If the typical event sequence of a process is interrupted by other events (e.g. a customer reneging, a failure) the process is blocked and has to use exception handling to tackle it.

Now

Processes (or asynchronous tasks in general) transfer IO-operations with a now! call to the (master) clock so that they get executed at the current clock time. As a convenience you can print directly to the clock.

DiscreteEvents.now! — Function.

now!(clk::Clock, ex::A) where {A<:Action}

Transfer an IO-operation ex to the master clock (on thread 1). The clock executes it before proceeding to the next time step.

source

Base.print — Method.

print(clk::Clock, [io::IO], xs...)

Let the clock clk print out xs to io (or to stdout).

source

Base.println — Method.

println(clk::Clock, [io::IO], xs...)

Let the clock clk print out xs to io (or to stdout) followed by a newline.

source