Internals

Internals

Module

Simulate.Simulate — Module.

Simulate

A Julia package for discrete event simulation based on state machines.

source

The module contains two main types: Clock and Logger. Both are implemented as state machines. The implementation functions and types are not exported. The exported functions documented above under Usage are commands to the internal state machines.

State machines

We have some definitions for them to work.

Simulate.SEngine — Type.

supertype for state machines in Sim.jl

source

States

Defined states for state machines.

Simulate.SState — Type.

supertype for states

source

Simulate.Undefined — Type.

a state machine is undefined (after creation)

source

Simulate.Idle — Type.

a state machine is idle

source

Simulate.Empty — Type.

a state machine is empty

source

Simulate.Busy — Type.

a state machine is busy

source

Simulate.Halted — Type.

a state machine is halted

source

Events

Defined events.

Simulate.SEvent — Type.

supertype for events

source

Simulate.Init — Type.

Init(info): Init event with some info.

source

Simulate.Setup — Type.

Setup(vars::Array{Symbol,1}, scope::Module): setup a logger with some info.

source

Simulate.Switch — Type.

Switch(to): switch to some other mode

source

Simulate.Log — Type.

Log(): record command for logging

source

Simulate.Step — Type.

Step(): command

source

Simulate.Run — Type.

Run(): command

source

Simulate.Start — Type.

Start(): command

source

Simulate.Stop — Type.

Stop(): command

source

Simulate.Resume — Type.

Resume(): command

source

Simulate.Clear — Type.

Clear(): command

source

Transition functions

In state machines transitions occur depending on states and events. The different transitions are described through different methods of the step!-function.

Simulate.step! — Function.

step!(A::SEngine, q::SState, σ::SEvent)

Default transition for clock and logger.

This is called if no otherwise defined transition occurs.

Arguments

A::SEngine: state machine for which a transition is called

q::SState: state of the state machine

σ::SEvent: event, triggering the transition

source

step!(sim::Clock, ::Undefined, ::Init)

initialize a clock.

source

step!(sim::Clock, ::Undefined, σ::Union{Step,Run})

if uninitialized, initialize and then Step or Run.

source

step!(sim::Clock, ::Union{Idle,Busy,Halted}, ::Step)

step forward to next tick or scheduled event. At a tick evaluate all sampling expressions, or, if an event is encountered evaluate the event expression.

The internal clock times sim.tev and sim.tsa must always be set to be at least sim.time.

source

step!(sim::Clock, ::Idle, σ::Run)

Run a simulation for a given duration. The duration is given with Run(duration). Call scheduled events and evaluate sampling expressions at each tick in that timeframe.

source

step!(sim::Clock, ::Busy, ::Stop)

Stop the clock.

source

step!(sim::Clock, ::Halted, ::Resume)

Resume a halted clock.

source

step!(sim::Clock, q::SState, σ::SEvent)

catch all step!-function.

source

step!(A::Logger, ::Undefined, σ::Init)

Initialize a logger.

source

step!(A::Logger, ::Empty, σ::Setup)

Setup a logger with logging variables. They are given by Setup(vars, scope).

source

step!(A::Logger, ::Idle, ::Clear)

Clear the last record and the data table of a logger.

source

step!(A::Logger, ::Idle, σ::Log)

Logging event.

source

step!(A::Logger, ::Idle, σ::Switch)

Switch the operating mode of a logger by Switch(to).

to = 0: no output, to = 1: print, `to = 2: store in log table"

source

Other internal types and functions

Simulate.SimEvent — Type.

SimEvent(expr::Expr, scope::Module, t::Float64, Δt::Float64)

Create a simulation event: an expression to be executed at event time.

Arguments

expr::Expr: expression to be evaluated at event time

scope::Module: evaluation scope

t::Float64: event time

Δt::Float64: repeat rate with which the event gets repeated

source

Simulate.Sample — Type.

Sample(ex::Union{Expr, SimFunction}, scope::Module)

Create a sampling expression.

Arguments

ex::{Expr, SimFunction}: expression or function to be called at sample time

scope::Module: evaluation scope

source

Simulate.simExec — Function.

simExec(ex::Union{Expr,SimFunction}, m::Module=Main)

evaluate an expression or execute a SimFunction.

source

Simulate.nextevent — Function.

nextevent(sim::Clock)

Return the next scheduled event.

source

Simulate.nextevtime — Function.

nextevtime(sim::Clock)

Return the internal time (unitless) of next scheduled event.

source

Simulate.checktime — Function.

checktime(sim::Clock, t::Number)::Float64

check t given according to clock settings and return value

source

Simulate.setTimes — Function.

setTimes(sim::Clock)

set clock times for next event or sampling action. The internal clock times sim.tev and sim.tsa must always be set to be at least sim.time.

source

Simulate.startup! — Function.

startup!(p::SimProcess)

Start a SimProcess as a task in a loop.

source

Simulate.loop — Function.

loop(p::SimProcess)

Put a SimProcess in a loop, which can be broken by a SimException.

source