Internals

Internals

The following types are handled internally by DiscreteEvents.jl. But they maybe interesting for analyzing and debugging clocks and event schedules.

Events

DiscreteEvents.AbstractEvent — Type.

supertype for events

source

DiscreteEvents.DiscreteEvent — Type.

DiscreteEvent{T<:Action} <: AbstractEvent

A discrete event is an Action to be executed at an event time.

Arguments, fields

ex::T: a function or an expression or a tuple of them,

t::Float64: event time,

Δt::Float64: repeat rate with for repeating events.

source

DiscreteEvents.DiscreteCond — Type.

DiscreteCond{S<:Action, T<:Action} <: AbstractEvent

A condition to be evaluated repeatedly with expressions or functions to be executed if conditions are met.

Arguments, fields

cond::S: a conditional function or an expression or a tuple of them (conditions must evaluate to Bool),

ex::T: a function or an expression or a tuple of them to be executed if conditions are met,

source

DiscreteEvents.Sample — Type.

Sample{T<:Action} <: AbstractEvent

Sampling actions are executed at sampling time.

Arguments, fields

ex<:Action: an Action to be executed at sample time.

source

DiscreteEvents.ClockException — Type.

ClockException(ev::ClockEvent, value=nothing)

Define a ClockException, which can be thrown to processes.

Arguments, fields

ev::ClockEvent: delivers an event to the interrupted task

value=nothing: deliver some other value

source

Clock

ActiveClocks are internal since the should not be setup explicitly.

DiscreteEvents.AbstractClock — Type.

supertype for clocks in DiscreteEvents.jl

source

DiscreteEvents.ActiveClock — Type.

ActiveClock{E <: ClockEvent} <: AbstractClock

A thread specific clock which can be operated via a channel.

Fields

clock::Clock: the thread specific clock,

master::Ref{Clock}: a pointer to the master clock (on thread 1),

cmd::Channel{E}: the command channel from master,

ans::Channel{E}: the response channel to master.

id::Int: the id in master's ac array,

thread::Int: the thread, the active clock runs on.

Note

You should not setup an ActiveClock explicitly. Rather this is done implicitly by fork!ing a Clock to other available threads or directly with PClock. It then can be accessed via pclock as in the following example.

Note

Directly accessing the clock-field of parallel ActiveClocks in simulations is possible but not recommended since it breaks parallel operation. The right way is to pass event!s to the ActiveClock-variable. The communication then happens over the channel to the ActiveClock as it should be.

Example

julia> using DiscreteEvents

julia> clk = Clock()
Clock 0, thrd 1 (+ 0 ac): state=DiscreteEvents.Undefined(), t=0.0 , Δt=0.01 , prc:0
  scheduled ev:0, cev:0, sampl:0

julia> fork!(clk)

julia> clk    #  ⬇ here you got 3 parallel active clocks
Clock 0, thrd 1 (+ 3 ac): state=DiscreteEvents.Undefined(), t=0.0 , Δt=0.01 , prc:0
  scheduled ev:0, cev:0, sampl:0

julia> clk = PClock()
Clock 0, thrd 1 (+ 3 ac): state=DiscreteEvents.Undefined(), t=0.0 , Δt=0.01 , prc:0
  scheduled ev:0, cev:0, sampl:0

julia> ac1 = pclock(clk, 1)    # get access to the 1st active clock (on thread 2)
Active clock 1 on thread 2: state=DiscreteEvents.Idle(), t=0.0 , Δt=0.01 , prc:0
   scheduled ev:0, cev:0, sampl:0

source

DiscreteEvents.Schedule — Type.

Schedule()

A Schedule contains events, conditional events and sampling functions to be executed or evaluated on the clock's time line.

Fields

events::PriorityQueue{DiscreteEvent,Float64}: scheduled events,

cevents::Array{DiscreteCond,1}: conditional events to evaluate at each tick,

samples::Array{Sample,1}: sampling expressions to evaluate at each tick,

source

DiscreteEvents.ClockChannel — Type.

ClockChannel

Provide a channel to an active clock or a real time clock.

Fields

ref::Ref{Task}: a pointer to an active clock,

ch::Channel: a communication channel to an active clock,

id::Int: the thread id of the active clock.

done::Bool: flag indicating if the active clock has completed its cycle.

source