Internals

Internals

Module

Sim.Sim — Module.

Sim

A Julia package for discrete event simulation based on state machines.

source

The module contains two main types: Clock and Logger. Both are implemented as state machines. The implementation functions and types are not exported. The exported functions documented above under Usage are commands to the internal state machines.

State machines

We have some definitions for them to work.

Sim.SEngine — Type.

supertype for state machines in Sim.jl

source

States

Defined states for state machines.

Sim.SState — Type.

supertype for states

source

Sim.Undefined — Type.

a state machine is undefined (after creation)

source

Sim.Idle — Type.

a state machine is idle

source

Sim.Empty — Type.

a state machine is empty

source

Sim.Busy — Type.

a state machine is busy

source

Sim.Halted — Type.

a state machine is halted

source

Events

Defined events.

Sim.SEvent — Type.

supertype for events

source

Sim.Init — Type.

Init(info): Init event with some info.

source

Sim.Setup — Type.

Setup(vars::Array{Symbol,1}, scope::Module): setup a logger with some info.

source

Sim.Switch — Type.

Switch(to): switch to some other mode

source

Sim.Log — Type.

Log(): record command for logging

source

Sim.Step — Type.

Step(): command

source

Sim.Run — Type.

Run(): command

source

Sim.Start — Type.

Start(): command

source

Sim.Stop — Type.

Stop(): command

source

Sim.Resume — Type.

Resume(): command

source

Sim.Clear — Type.

Clear(): command

source

Transition functions

In state machines transitions occur depending on states and events. The different transitions are described through different methods of the step!-function.

Sim.step! — Function.

step!(A::SEngine, q::SState, σ::SEvent)

Default transition for clock and logger.

This is called if no otherwise defined transition occurs.

Arguments

A::SEngine: state machine for which a transition is called

q::SState: state of the state machine

σ::SEvent: event, triggering the transition

source

step!(sim::Clock, ::Undefined, ::Init)

initialize a clock.

source

step!(sim::Clock, ::Undefined, σ::Union{Step,Run})

if uninitialized, initialize and then Step or Run.

source

step!(sim::Clock, ::Union{Idle,Busy,Halted}, ::Step)

step forward to next tick or scheduled event.

At a tick evaluate all sampling expressions, or, if an event is encountered evaluate the event expression.

source

step!(sim::Clock, ::Idle, σ::Run)

Run a simulation for a given duration.

The duration is given with Run(duration). Call scheduled events and evaluate sampling expressions at each tick in that timeframe.

source

step!(sim::Clock, ::Busy, ::Stop)

Stop the clock.

source

step!(sim::Clock, ::Halted, ::Resume)

Resume a halted clock.

source

step!(sim::Clock, q::SState, σ::SEvent)

catch all step!-function.

source

step!(A::Logger, ::Undefined, σ::Init)

Initialize a logger.

source

step!(A::Logger, ::Empty, σ::Setup)

Setup a logger with logging variables. They are given by Setup(vars, scope).

source

step!(A::Logger, ::Idle, ::Clear)

Clear the last record and the data table of a logger.

source

step!(A::Logger, ::Idle, σ::Log)

Logging event.

source

step!(A::Logger, ::Idle, σ::Switch)

Switch the operating mode of a logger by Switch(to).

to = 0: no output, to = 1: print, `to = 2: store in log table"

source

Other internal types and functions

Sim.SimEvent — Type.

SimEvent(expr::Expr, scope::Module, t::Time, Δt::Time)

Create a simulation event: an expression to be executed at event time.

Arguments

expr::Expr: expression to be evaluated at event time

scope::Module: evaluation scope

t::Float64: event time

Δt::Float64: repeat rate with which the event gets repeated

source

Sim.Sample — Type.

Sample(ex::Union{Expr, SimFunction}, scope::Module)

Create a sampling expression.

Arguments

ex::{Expr, SimFunction}: expression or function to be called at sample time

scope::Module: evaluation scope

source

Sim.simExec — Function.

simExec(ex::Union{Expr,SimFunction}, m::Module=Main)

evaluate an expression or execute a SimFunction.

source

Sim.nextevent — Function.

nextevent(sim::Clock)

Return the next scheduled event.

source

Sim.nextevtime — Function.

nextevtime(sim::Clock)

Return the time of next scheduled event.

source