Processes

Processes are typical event sequences. They get implemented in a function, run as asynchronous tasks in a loop and get registered to a clock.

DiscreteEvents.Prc — Type

Prc(id, f, arg...; kw...)

Prepare a function to run as a process registered to a clock.

Arguments, fields

id: some unique identification for registration,

f::Function: a function f(clk, arg...; kw...), must take clk (a Clock) as its first argument,

arg...: further arguments to f

kw...: keyword arguments to f

Fields

task::Union{Task,Nothing}: a task structure,

clk::Union{AbstractClock,Nothing}: clock where the process is registered,

state::ClockState: process state,

DiscreteEvents.process! — Function

process!([clk], prc, cycles; <keyword arguments>)

Register a Prc to a clock, start an asynchronous task executing the process function in a loop and return the id it was registered with. It can then be found under clk.processes[id].

Arguments

c<:AbstractClock: if not provided, the process runs under 𝐶,

prc::Prc: it contains a function and its arguments,

cycles<:Number=Inf: number of loop cycles the process should run,

Keyword arguments

cid::Int=clk.id: if cid ≠ clk.id, assign the event to the parallel clock with id == cid. This overrides spawn,

spawn::Bool=false: if true, the process may be scheduled on another thread in parallel and registered to the thread specific clock.

Processes must yield!

A process has to give control back to Julia by e.g. doing a take!(input) or by calling delay! or wait!. Otherwise it will starve everything else!

Delay and wait …

In order to implement a process (an event sequence) functions can call delay! or wait! on the clock or take! and put! on channels. They are then suspended until a given time or until certain conditions are met or requested resources are available.

DiscreteEvents.delay! — Function

delay!(clk, Δt)
delay!(clk, T, t)

Delay (suspend) a process for a time interval Δt on the clock clk.

Arguments

clk::Clock,

Δt: time interval, Number or Distribution,

T::Timing: only until is accepted,

t: time delay, Number or Distribution.

DiscreteEvents.wait! — Function

wait!(clk, cond)

Delay (suspend) a process on a clock clk until a condition has become true.

Arguments

clk::Clock,

cond<:Action: a condition, true if all expressions or functions therein return true.

Interrupts

If other events (customers reneging, failures) interrupt the typical event sequence of a process, it is blocked and not ready to respond. Processes therefore must use exception handling to handle unusual events.

DiscreteEvents.PrcException — Type

PrcException(event, value)

An exception to be thrown at processes.

Arguments, fields

event: deliver an event to the interrupted task

value: deliver some other value

DiscreteEvents.interrupt! — Function

interrupt!(p::Prc, ev, value)

Interrupt a Prc by throwing a PrcException to it.

An example at DiscreteEventsCompanion illustrates how to handle interrupts to a process. Things can get messy quickly if there are several unusual events which have to be handled in a process.

Now

Processes (or asynchronous tasks in general) transfer IO-operations with a now! call to the (master) clock so that they get executed at the current clock time. As a convenience you can print directly to the clock.

DiscreteEvents.now! — Function

now!(clk::Clock, ex::A) where {A<:Action}

Transfer an IO-operation ex to the master clock (on thread 1). The clock executes it before proceeding to the next time step.

Base.print — Method

print(clk::Clock, [io::IO], x, xs...)

Create a now! event to a busy clock clk to print(x, xs...) to io (or to stdout).

If the clock is not busy, x and xs... are printed as usual. print(clk) still prints repr(clk).

Base.println — Method

println(clk::Clock, [io::IO], xs...)

Create a now! event to a busy clock clk to println(x, xs...) to io (or to stdout).

If the clock is not busy, x and xs... are printed as usual. println(clk) still prints repr(clk).