Usage

User guide

Sim.jl is not yet an registered package and is installed with

pkg> add("https://github.com/pbayer/jl")

The package is then loaded with

julia> using Sim

Modeling and simulation

A Clock allows the definition of scheduled events and sampling actions, which occur at predefined ticks. When we run! the Clock, it fires the events at their scheduled times and executes the sampling actions at each tick.

Types

Sim.Clock — Type.

Clock(Δt::Number=0; t0::Number=0)

Create a new simulation clock.

Arguments

Δt::Number=0: time increment

t0::Number=0: start time for simulation.

If no Δt is given, the simulation doesn't tick, but jumps from event to event. Δt can be set later with sample_time!.

source

Sim.Timing — Type.

Timing

Enumeration type for scheduling events and timed conditions:

at: schedule an event at agiven time

after: schedule an event a given time after current time

every: schedule an event every given time from now

before: a timed condition is true before a given time.

source

Functions

Sim.now — Method.

now(sim::Clock)

Return the current simulation time.

source

Sim.sample_time! — Method.

sample_time!(sim::Clock, Δt::Number)

set the clock's sampling time from now(sim).

Arguments

sim::Clock

Δt::Number: sample rate, time interval for sampling

source

Sim.event! — Function.

    event!(sim::Clock, expr::Expr, t::Number; scope::Module=Main, cycle::Number=0.0)::Float64
    event!(sim::Clock, expr::Expr, T::Timing, t::Number; scope::Module=Main)::Float64

Schedule an expression for execution at a given simulation time.

Arguments

sim::Clock: simulation clock

expr::Expr: an expression

t::Float64: simulation time

scope::Module=Main: scope for the expression to be evaluated

cycle::Float64=0.0: repeat cycle time for the event

T::Timing: a timing, at, after or every (before behaves like at)

returns

Scheduled simulation time for that event.

May return a time t > at from repeated applications of nextfloat(at) if there were yet events scheduled for that time.

source

Sim.sample! — Function.

sample!(sim::Clock, expr::Expr; scope::Module=Main)

enqueue an expression for sampling.

Arguments

sim::Clock

expr::Expr: a Julia expression

scope::Module=Main: optional, a scope for the expression to be evaluated in

source

Sim.incr! — Function.

incr!(sim::Clock)

Take one simulation step, execute the next tick or event.

source

Sim.run! — Function.

run!(sim::Clock, duration::Number)

Run a simulation for a given duration.

Call scheduled events and evaluate sampling expressions at each tick in that timeframe.

source

Sim.stop! — Function.

stop!(sim::Clock)

Stop a running simulation.

source

Sim.resume! — Function.

resume!(sim::Clock)

Resume a halted simulation.

source

Logging

A Logger allows to register variables and to record their states on demand. The last record is stored in the logging variable. According to the Logger's state it can be printed or stored in a table.

julia> sim = Clock(); # create a clock

julia> l = Logger(); # create a logging variable

julia> init!(l, sim); # initialize the logger

julia> setup!(l, [:a, :b, :c]); # register the variables a, b, c for logging

julia> record!(l); # record the variables with the current clock time

Types

Sim.Logger — Type.

Logger()

Setup and return a logging variable.

source

Functions

Sim.init! — Function.

init!(sim::Clock)

initialize a clock.

source

init!(L::Logger, sim::Clock)

Initialize a Logger.

source

Sim.setup! — Function.

setup!(L::Logger, vars::Array{Symbol})

Setup a logger with logging variables.

Arguments

L::Logger

vars::Array{Symbol}: An array of symbols, e.g. of global variables

source

Sim.switch! — Function.

switch!(L::Logger, to::Number=0)

Switch the operating mode of a logger.

to = 0: no output, to = 1: print, `to = 2: store in log table"

source

Sim.record! — Function.

record!(L::Logger)

record the logging variables with the current operating mode.

source

Sim.clear! — Function.

clear!(L::Logger)

clear the loggers last record and data table.

source

The recorded values can be accessed with

julia> l.last
(time = 0.0, a = 0, b = 0, c = 0)

julia> l.df
5×4 DataFrames.DataFrame
│ Row │ time    │ a     │ b     │ c     │
│     │ Float64 │ Int64 │ Int64 │ Int64 │
├─────┼─────────┼───────┼───────┼───────┤
│ 1   │ 0.0     │ 10    │ 100   │ 1331  │
│ 2   │ 0.0     │ 10    │ 100   │ 1728  │
│ 3   │ 0.0     │ 10    │ 100   │ 2197  │
│ 4   │ 0.0     │ 10    │ 100   │ 2744  │
│ 5   │ 0.0     │ 10    │ 100   │ 3375  │