Usage

User guide

Sim.jl is not yet an registered package and is installed with

pkg> add("https://github.com/pbayer/jl")

The package is then loaded with

julia> using Sim

Modeling and simulation

A virtual Clock allows to schedule Julia functions or expressions as timed events or as sampling actions, which occur at predefined clock ticks. When we run the Clock, it fires the events at their scheduled times and executes the sampling actions at each tick.

Silly example

julia> m = @__MODULE__ # catching the current module is needed for documentation
Main.ex-usage

julia> using Printf

julia> comm = ["Hi, nice to meet you!", "How are you?", "Have a nice day!"];

julia> greet(name, n) =  @printf("%5.2f s, %s: %s\n", τ(), name, comm[n])
greet (generic function with 1 method)

julia> function foo(n) # 1st passerby
           greet("Foo", n)
           event!(Τ, :(bar($n)), after, 2*rand(), scope = m)
       end
foo (generic function with 1 method)

julia> function bar(n) # 2nd passerby
           greet("Bar", n)
           if n < 3
              event!(Τ, :(foo($n+1)), after, 2*rand(), scope = m)
           else
              println("bye bye")
           end
       end
bar (generic function with 1 method)

julia> event!(Τ, :(foo(1)), at, 10*rand(), scope = m); # create an event for a start

julia> run!(Τ, 20) # and run the simulation
 9.99 s, Foo: Hi, nice to meet you!
11.36 s, Bar: Hi, nice to meet you!
12.46 s, Foo: How are you?
13.50 s, Bar: How are you?
14.07 s, Foo: Have a nice day!
14.36 s, Bar: Have a nice day!
bye bye
Finished: 6 events, simulation time: 20.0

Types and constants

Sim.Time — Type.

Time

Time is a number in Sim.jl

source

Sim.Clock — Type.

Clock(Δt::Number=0; t0::Number=0)

Create a new simulation clock.

Arguments

Δt::Number=0: time increment

t0::Number=0: start time for simulation.

If no Δt is given, the simulation doesn't tick, but jumps from event to event. Δt can be set later with sample_time!.

source

Sim.Timing — Type.

Timing

Enumeration type for scheduling events and timed conditions:

at: schedule an event at a given time

after: schedule an event a given time after current time

every: schedule an event every given time from now on

before: a timed condition is true before a given time.

source

Sim.SimFunction — Type.

SimFunction(func::Function, arg...; kw...)

Type for preparing a function as an event to a simulation.

Arguments

func::Function: function to be executed at a later simulation time

arg...: arguments to the function

kw...: keyword arguments

Be aware that, if the variables stored in a SimFunction are composite types, they can change until they are evaluated later by func. But that's the nature of simulation.

Example

julia> using Sim

julia> f(a,b,c; d=4, e=5) = a+b+c+d+e  # define a function
f (generic function with 1 method)

julia> sf = SimFunction(f, 10, 20, 30, d=14, e=15)  # store it as SimFunction
SimFunction(f, (10, 20, 30), Base.Iterators.Pairs(:d => 14,:e => 15))

julia> sf.func(sf.arg...; sf.kw...)  # and it can be executed later
89

julia> d = Dict(:a => 1, :b => 2) # now we set up a dictionary
Dict{Symbol,Int64} with 2 entries:
  :a => 1
  :b => 2

julia> f(t) = t[:a] + t[:b] # and a function adding :a and :b
f (generic function with 2 methods)

julia> f(d)  # our add function gives 3
3

julia> ff = SimFunction(f, d)   # we set up a SimFunction
SimFunction(f, (Dict(:a => 1,:b => 2),), Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}}())

julia> d[:a] = 10  # later somehow we need to change d
10

julia> ff  # our SimFunction ff has changed too
SimFunction(f, (Dict(:a => 10,:b => 2),), Base.Iterators.Pairs{Union{},Union{},Tuple{},NamedTuple{(),Tuple{}}}())

julia> ff.func(ff.arg...; ff.kw...)  # and calling it gives a different result
12

source

Central time

Sim.Τ — Constant.

Τ
Tau

Τ (\Tau+Tab), Tau is the central Clock()-variable.

Examples

julia> using Sim

julia> Τ  # central clock
Clock(Sim.Idle(), 0, DataStructures.PriorityQueue{Sim.SimEvent,Float64,Base.Order.ForwardOrdering}(), 0, 0, 0, 0, Sim.Sample[], 0)
julia> Tau  # alias
Clock(Sim.Idle(), 0, DataStructures.PriorityQueue{Sim.SimEvent,Float64,Base.Order.ForwardOrdering}(), 0, 0, 0, 0, Sim.Sample[], 0)
julia> Τ.time
0

source

Functions

Sim.τ — Function.

τ(sim::Clock=Τ)
tau(sim::Clock=Tau)

Return the current simulation time (τ= au+Tab).

Examples

julia> using Sim

julia> τ() # gives the central time
0
julia> tau() # alias, gives the central time
0

source

Sim.sample_time! — Function.

sample_time!(sim::Clock, Δt::Time)

set the clock's sampling time from au(sim).

Arguments

sim::Clock

Δt::Number: sample rate, time interval for sampling

source

Sim.event! — Function.

event!(sim::Clock, ex::Union{Expr, SimFunction}, t::Number; scope::Module=Main, cycle::Number=0.0)::Float64
event!(sim::Clock, ex::Union{Expr, SimFunction}, T::Timing, t::Number; scope::Module=Main)::Float64

Schedule a function or expression for a given simulation time.

Arguments

sim::Clock: simulation clock

ex::{Expr, SimFunction}: an expression or SimFunction

t::Time: simulation time

scope::Module=Main: scope for the expression to be evaluated

cycle::Time=0.0: repeat cycle time for the event

T::Timing: a timing, at, after or every (before behaves like at)

returns

Scheduled simulation time for that event.

May return a time t > at from repeated applications of nextfloat(at) if there were yet events scheduled for that time.

source

Sim.sample! — Function.

sample!(sim::Clock, ex::Union{Expr, SimFunction}; scope::Module=Main)

enqueue an expression for sampling.

Arguments

sim::Clock

ex::Union{Expr, SimFunction}: an expression or function

scope::Module=Main: optional, a scope for the expression to be evaluated in

source

Sim.incr! — Function.

incr!(sim::Clock)

Take one simulation step, execute the next tick or event.

source

Sim.run! — Function.

run!(sim::Clock, duration::Number)

Run a simulation for a given duration.

Call scheduled events and evaluate sampling expressions at each tick in that timeframe.

source

Sim.stop! — Function.

stop!(sim::Clock)

Stop a running simulation.

source

Sim.resume! — Function.

resume!(sim::Clock)

Resume a halted simulation.

source

Sim.sync! — Function.

sync!(sim::Clock, to::Clock=Τ)

Force a synchronization of two clocks. Change all registered times of sim accordingly.

source

Sim.reset! — Function.

reset!(sim::Clock, Δt::Number=0; t0::Time=0, hard::Bool=true)

reset a clock

Arguments

sim::Clock

Δt::Number=0: time increment

t0::Time=0: start time

hard::Bool=true: time is reset, all scheduled events and sampling are

deleted. If hard=false, then only time is reset, event and sampling times are adjusted accordingly.

source

Logging

A Logger allows to register variables and to record their states on demand. The last record is stored in the logging variable. According to the Logger's state it can be printed or stored in a table.

Example

julia> sim = Clock(); # create a clock

julia> l = Logger(); # create a logging variable

julia> init!(l, sim); # initialize the logger

julia> (a, b, c) = 1, 1, 1 # create some variables
(1, 1, 1)

julia> setup!(l, [:a, :b, :c], scope = m); # register them for logging

julia> record!(l) # record the variables with the current clock time

julia> l.last # show the last record
(time = 0, a = 1, b = 1, c = 1)

julia> function f()  # a function for increasing and recording the variables
         global a += 1
         global b = a^2
         global c = a^3
         record!(l)
       end
f (generic function with 1 method)

julia> switch!(l, 1); # switch logger to printing

julia> f() # increase and record the variables
(time = 0, a = 2, b = 4, c = 8)

julia> switch!(l, 2); # switch logger to storing in data table

julia> for i in 1:10 # create some events
           event!(sim, :(f()), i, scope = m)
       end

julia> run!(sim, 10) # run a simulation
Finished: 10 events, simulation time: 10.0

julia> l.df # view the recorded values
10×4 DataFrames.DataFrame
│ Row │ time    │ a     │ b     │ c     │
│     │ Float64 │ Int64 │ Int64 │ Int64 │
├─────┼─────────┼───────┼───────┼───────┤
│ 1   │ 1.0     │ 3     │ 9     │ 27    │
│ 2   │ 2.0     │ 4     │ 16    │ 64    │
│ 3   │ 3.0     │ 5     │ 25    │ 125   │
│ 4   │ 4.0     │ 6     │ 36    │ 216   │
│ 5   │ 5.0     │ 7     │ 49    │ 343   │
│ 6   │ 6.0     │ 8     │ 64    │ 512   │
│ 7   │ 7.0     │ 9     │ 81    │ 729   │
│ 8   │ 8.0     │ 10    │ 100   │ 1000  │
│ 9   │ 9.0     │ 11    │ 121   │ 1331  │
│ 10  │ 10.0    │ 12    │ 144   │ 1728  │

Types

Sim.Logger — Type.

Logger()

Setup and return a logging variable.

source

Functions

Sim.init! — Function.

init!(sim::Clock)

initialize a clock.

source

init!(L::Logger, sim::Clock)

Initialize a Logger.

source

Sim.setup! — Function.

setup!(L::Logger, vars::Array{Symbol})

Setup a logger with logging variables.

Arguments

L::Logger

vars::Array{Symbol}: An array of symbols, e.g. of global variables

scope::Module = Main: Scope in which to evaluate the variables

source

Sim.switch! — Function.

switch!(L::Logger, to::Number=0)

Switch the operating mode of a logger.

to = 0: no output, to = 1: print, `to = 2: store in log table"

source

Sim.record! — Function.

record!(L::Logger)

record the logging variables with the current operating mode.

source

Sim.clear! — Function.

clear!(L::Logger)

clear the loggers last record and data table.

source