Usage

User guide

Sim.jl is not yet an registered package and is installed with

pkg> add("https://github.com/pbayer/jl")

The package is then loaded with

julia> using Sim

Modeling and simulation

A virtual Clock allows to schedule Julia expressions as timed events or as sampling actions, which occur at predefined clock ticks. When we run the Clock, it fires the events at their scheduled times and executes the sampling actions at each tick.

Silly example

julia> m = @__MODULE__ # catching the current module is needed for documentation
Main.ex-usage

julia> using Printf

julia> sim = Clock(); # create a clock

julia> comm = ["Hi, nice to meet you!", "How are you?", "Have a nice day!"];

julia> greet(name, n) =  @printf("%5.2f s, %s: %s\n", now(sim), name, comm[n])
greet (generic function with 1 method)

julia> function foo(n) # 1st passerby
           greet("Foo", n)
           event!(sim, :(bar($n)), after, 2*rand(), scope = m)
       end
foo (generic function with 1 method)

julia> function bar(n) # 2nd passerby
           greet("Bar", n)
           if n < 3
              event!(sim, :(foo($n+1)), after, 2*rand(), scope = m)
           else
              println("bye bye")
           end
       end
bar (generic function with 1 method)

julia> event!(sim, :(foo(1)), at, 10*rand(), scope = m); # create an event for a start

julia> run!(sim, 20) # and run the simulation
 9.91 s, Foo: Hi, nice to meet you!
10.33 s, Bar: Hi, nice to meet you!
12.23 s, Foo: How are you?
13.47 s, Bar: How are you?
15.05 s, Foo: Have a nice day!
16.55 s, Bar: Have a nice day!
bye bye
Finished: 6 events, simulation time: 20.0

Types

Sim.Clock — Type.

Clock(Δt::Number=0; t0::Number=0)

Create a new simulation clock.

Arguments

Δt::Number=0: time increment

t0::Number=0: start time for simulation.

If no Δt is given, the simulation doesn't tick, but jumps from event to event. Δt can be set later with sample_time!.

source

Sim.Timing — Type.

Timing

Enumeration type for scheduling events and timed conditions:

at: schedule an event at agiven time

after: schedule an event a given time after current time

every: schedule an event every given time from now

before: a timed condition is true before a given time.

source

Functions

Sim.now — Function.

now(sim::Clock)

Return the current simulation time.

source

Sim.sample_time! — Function.

sample_time!(sim::Clock, Δt::Number)

set the clock's sampling time from now(sim).

Arguments

sim::Clock

Δt::Number: sample rate, time interval for sampling

source

Sim.event! — Function.

    event!(sim::Clock, expr::Expr, t::Number; scope::Module=Main, cycle::Number=0.0)::Float64
    event!(sim::Clock, expr::Expr, T::Timing, t::Number; scope::Module=Main)::Float64

Schedule an expression for execution at a given simulation time.

Arguments

sim::Clock: simulation clock

expr::Expr: an expression

t::Float64: simulation time

scope::Module=Main: scope for the expression to be evaluated

cycle::Float64=0.0: repeat cycle time for the event

T::Timing: a timing, at, after or every (before behaves like at)

returns

Scheduled simulation time for that event.

May return a time t > at from repeated applications of nextfloat(at) if there were yet events scheduled for that time.

source

Sim.sample! — Function.

sample!(sim::Clock, expr::Expr; scope::Module=Main)

enqueue an expression for sampling.

Arguments

sim::Clock

expr::Expr: a Julia expression

scope::Module=Main: optional, a scope for the expression to be evaluated in

source

Sim.incr! — Function.

incr!(sim::Clock)

Take one simulation step, execute the next tick or event.

source

Sim.run! — Function.

run!(sim::Clock, duration::Number)

Run a simulation for a given duration.

Call scheduled events and evaluate sampling expressions at each tick in that timeframe.

source

Sim.stop! — Function.

stop!(sim::Clock)

Stop a running simulation.

source

Sim.resume! — Function.

resume!(sim::Clock)

Resume a halted simulation.

source

Logging

A Logger allows to register variables and to record their states on demand. The last record is stored in the logging variable. According to the Logger's state it can be printed or stored in a table.

Example

julia> sim = Clock(); # create a clock

julia> l = Logger(); # create a logging variable

julia> init!(l, sim); # initialize the logger

julia> (a, b, c) = 1, 1, 1 # create some variables
(1, 1, 1)

julia> setup!(l, [:a, :b, :c], scope = m); # register them for logging

julia> record!(l) # record the variables with the current clock time

julia> l.last # show the last record
(time = 0.0, a = 1, b = 1, c = 1)

julia> function f()  # a function for increasing and recording the variables
         global a += 1
         global b = a^2
         global c = a^3
         record!(l)
       end
f (generic function with 1 method)

julia> switch!(l, 1); # switch logger to printing

julia> f() # increase and record the variables
(time = 0.0, a = 2, b = 4, c = 8)

julia> switch!(l, 2); # switch logger to storing in data table

julia> for i in 1:10 # create some events
           event!(sim, :(f()), i, scope = m)
       end

julia> run!(sim, 10) # run a simulation
Finished: 10 events, simulation time: 10.0

julia> l.df # view the recorded values
10×4 DataFrames.DataFrame
│ Row │ time    │ a     │ b     │ c     │
│     │ Float64 │ Int64 │ Int64 │ Int64 │
├─────┼─────────┼───────┼───────┼───────┤
│ 1   │ 1.0     │ 3     │ 9     │ 27    │
│ 2   │ 2.0     │ 4     │ 16    │ 64    │
│ 3   │ 3.0     │ 5     │ 25    │ 125   │
│ 4   │ 4.0     │ 6     │ 36    │ 216   │
│ 5   │ 5.0     │ 7     │ 49    │ 343   │
│ 6   │ 6.0     │ 8     │ 64    │ 512   │
│ 7   │ 7.0     │ 9     │ 81    │ 729   │
│ 8   │ 8.0     │ 10    │ 100   │ 1000  │
│ 9   │ 9.0     │ 11    │ 121   │ 1331  │
│ 10  │ 10.0    │ 12    │ 144   │ 1728  │

Types

Sim.Logger — Type.

Logger()

Setup and return a logging variable.

source

Functions

Sim.init! — Function.

init!(sim::Clock)

initialize a clock.

source

init!(L::Logger, sim::Clock)

Initialize a Logger.

source

Sim.setup! — Function.

setup!(L::Logger, vars::Array{Symbol})

Setup a logger with logging variables.

Arguments

L::Logger

vars::Array{Symbol}: An array of symbols, e.g. of global variables

scope::Module = Main: Scope in which to evaluate the variables

source

Sim.switch! — Function.

switch!(L::Logger, to::Number=0)

Switch the operating mode of a logger.

to = 0: no output, to = 1: print, `to = 2: store in log table"

source

Sim.record! — Function.

record!(L::Logger)

record the logging variables with the current operating mode.

source

Sim.clear! — Function.

clear!(L::Logger)

clear the loggers last record and data table.

source