User guide

User guide

Installation

DiscreteEvents.DiscreteEvents — Module.

DiscreteEvents

A Julia package for generating and simulating discrete events. It runs on Julia VERSION ≥ v"1.0". Multithreading requires VERSION ≥ v"1.3".

The current stable, registered version is installed with

pkg> add DiscreteEvents

The development version is installed with:

pkg> add("https://github.com/pbayer/DiscreteEvents.jl")

source

DiscreteEvents.version — Constant.

version

Gives the package version:

julia> using DiscreteEvents

julia> DiscreteEvents.version
v"0.3.0"

source

Clocks

Clock schedule and execute Julia functions or expressions as events at given times or under given conditions.

Clocks have virtual time and precede as fast as possible when they simulate chains of events. For parallel simulations they can control ActiveClocks on parallel threads.

RTClocks schedule and execute events on a real (system) time line.

Clocks have an ident number:

a master Clock on thread has id = 0,

worker ActiveClocks on parallel threads have id ≥ 1,

real time RTClocks have id ≤ -1.

DiscreteEvents.Clock — Type.

Clock(Δt::T=0.01; t0::U=0, unit::FreeUnits=NoUnits) where {T<:Number,U<:Number}

Create a new simulation clock.

Arguments

Δt::T=0.01: time increment for sampling. Δt can be set later with sample_time!.

t0::U=0: start time for simulation

unit::FreeUnits=NoUnits: clock time unit. Units can be set explicitely by setting e.g. unit=minute or implicitly by giving Δt as a time or else setting t0 to a time, e.g. t0=60s.

Fields

id::Int: clock ident number, 0: master clock, ≥ 1: parallel clock,

state::ClockState: clock state,

time::Float64: clock time,

unit::FreeUnits: time unit,

end_time::Float64: end time for simulation,

Δt::Float64: sampling time, timestep between ticks,

ac::Vector{ClockChannel}: channels to active clocks on parallel threads,

sc::Schedule: the clock Schedule (events, cond events and sampling),

processes::Dict{Any, Prc}: registered Prces,

tn::Float64: next timestep,

tev::Float64: next event time,

evcount::Int: event counter,

scount::Int: sample counter

Examples

julia> using DiscreteEvents, Unitful

julia> import Unitful: s, minute, hr

julia> c = Clock()                 # create a unitless clock (standard)
Clock 0, thread 1 (+ 0 ac): state=DiscreteEvents.Undefined(), t=0.0 , Δt=0.01 , prc:0
  scheduled ev:0, cev:0, sampl:0

julia> c1 = Clock(1s, unit=minute)  # create a clock with unit [minute]
Clock 0, thread 1 (+ 0 ac): state=DiscreteEvents.Undefined(), t=0.0 minute, Δt=0.01667 minute, prc:0
  scheduled ev:0, cev:0, sampl:0

julia> c2 = Clock(1s)               # create a clock with implicit unit [s]
Clock 0, thread 1 (+ 0 ac): state=DiscreteEvents.Undefined(), t=0.0 s, Δt=1.0 s, prc:0
  scheduled ev:0, cev:0, sampl:0

julia> c3 = Clock(t0=60s)           # another clock with implicit unit [s]
Clock 0, thread 1 (+ 0 ac): state=DiscreteEvents.Undefined(), t=60.0 s, Δt=0.01 s, prc:0
  scheduled ev:0, cev:0, sampl:0

julia> c4 = Clock(1s, t0=1hr)       # here Δt's unit [s] takes precedence
Clock 0, thread 1 (+ 0 ac): state=DiscreteEvents.Undefined(), t=3600.0 s, Δt=1.0 s, prc:0
  scheduled ev:0, cev:0, sampl:0

source

DiscreteEvents.RTClock — Type.

RTClock{E <: ClockEvent} <: AbstractClock

A real time clock checks every given period for scheduled events and executes them. It has a time in seconds since its start or last reset and uses system time for updating.

Real time clocks are controlled over channels. Multiple real time clocks can be setup with arbitrary periods (≥ 1 ms). Real time clocks should not be created directly but rather with CreateRTClock.

Fields

Timer::Timer: clock period in seconds, minimum is 0.001 (1 ms)

clock::Clock: clock work

cmd::Channel{T}: command channel to asynchronous clock

back::Channel{T}: back channel from async clock

id::Int: arbitrary id number

thread::Int: thread the async clock is living in

time::Float64: clock time since start in seconds

t0::Float64: system time at clock start in seconds

T::Float64: clock period in seconds

source

You can set time units and query the current clock time.

DiscreteEvents.setUnit! — Function.

setUnit!(clk::Clock, new::FreeUnits)

set a clock to a new time unit in Unitful. If necessary convert current clock times to the new unit.

Arguments

clk::Clock

new::FreeUnits: new is one of ms, s, minute or hr or another Unitful Time unit.

Examples

julia> using DiscreteEvents, Unitful

julia> import Unitful: Time, s, minute, hr

julia> c = Clock(t0=60)     # setup a new clock with t0=60
Clock 0, thread 1 (+ 0 ac): state=DiscreteEvents.Undefined(), t=60.0 , Δt=0.01 , prc:0
  scheduled ev:0, cev:0, sampl:0

julia> tau(c) # current time is 60.0 NoUnits
60.0

julia> setUnit!(c, s)       # set clock unit to Unitful.s
60.0 s

julia> tau(c) # current time is now 60.0 s
60.0 s

julia> setUnit!(c, minute)  # set clock unit to Unitful.minute
1.0 minute

julia> tau(c)               # current time is now 1.0 minute
1.0 minute

julia> isa(tau(c), Time)
true

julia> uconvert(s, tau(c))  # ... which can be converted to other time units
60.0 s

julia> tau(c).val           # it has a value of 1.0
1.0

julia> c.time               # internal clock time is set to 1.0 (a Float64)
1.0

julia> c.unit               # internal clock unit is set to Unitful.minute
minute

source

DiscreteEvents.tau — Function.

tau(clk::Clock=𝐶)

Return the current simulation time.

Examples

julia> using DiscreteEvents

julia> resetClock!(𝐶)
"clock reset to t₀=0.0, sampling rate Δt=0.01."
julia> tau() # gives the central time
0.0

source

There is a default clock 𝐶, which can be used for experimental work.

DiscreteEvents.𝐶 — Constant.

𝐶

𝐶 (𝐶 = \itC+tab) is the default simulation clock. If you do one simulation at a time, you can use it for time keeping.

Examples

julia> using DiscreteEvents

julia> resetClock!(𝐶)
"clock reset to t₀=0.0, sampling rate Δt=0.01."

julia> 𝐶  # default clock
Clock 0, thread 1 (+ 0 ac): state=DiscreteEvents.Idle(), t=0.0 , Δt=0.01 , prc:0
  scheduled ev:0, cev:0, sampl:0

source

You can create a clock with parallel active clocks on all available threads or fork existing clocks to other threads or collapse them if no longer needed. You can get direct access to parallel ActiveClocks and diagnose them.

DiscreteEvents.PClock — Function.

PClock(Δt::T=0.01; t0::U=0.0, unit::FreeUnits=NoUnits) where {T<:Number,U<:Number}

Setup a clock with parallel clocks on all available threads.

Arguments

Δt::T=0.01: time increment > 0. If given Δt ≤ 0, it gets set to 0.01.

t0::U=0.0: start time for simulation,

unit::FreeUnits=NoUnits: clock time unit for explicitl unit setting.

Note

Processes on multiple threads are possible in Julia ≥ 1.3 and with JULIA_NUM_THREADS > 1.

source

DiscreteEvents.fork! — Function.

fork!(master::Clock)

Establish copies of a master clock (thread 1) on all parallel threads.

source

DiscreteEvents.pclock — Function.

pclock(clk::Clock, id::Int ) :: C where {C<:AbstractClock}
pclock(ac::ActiveClock, id::Int ) :: C where {C<:AbstractClock}

Get a parallel clock to a given clock.

Arguments

master::Clock: a master clock or

ac::ActiveClock: an active clock,

id::Int=threadid(): thread id, defaults to the caller's current thread.

Returns

the master Clock if id==0,

a parallel ActiveClock else

source

DiscreteEvents.collapse! — Function.

collapse!(master::Clock)

Transfer the schedules of the parallel clocks to master and them stop them.

Note

If there are processes on other threads registered to parallel clocks, make sure that they aren't needed anymore before calling collapse. They are not transferred to and cannot be controlled by master.

source

DiscreteEvents.diagnose — Function.

diagnose(master::Clock, id::Int)

Return the stacktrace from parallel clock id.

source

DiscreteEvents.createRTClock — Function.

createRTClock(T::Float64, id::Int, thrd::Int=nthreads(); ch_size::Int=256)

Create, start and return a real time Clock.

The clock takes the current system time and starts to count in seconds with the given period T. Events or sampling functions can then be scheduled to it.

Arguments

T::Float64: period (clock resolution) in seconds, T ≥ 0.001

id::Int: clock identification number other than 0:(nthreads()-1)

thrd::Int=nthreads(): thread, the clock task should run in

ch_size::Int=256: clock communication channel size

source

DiscreteEvents.stopRTClock — Function.

stopRTClock(rtc::RTClock)

Stop a real time clock.

source

Events

Julia functions and expressions can be scheduled for execution

at given clock times and

under specified conditions.

DiscreteEvents.Action — Constant.

Action

An action is either a Function or an Expr or a Tuple of them. It can be scheduled in an event for later execution.

Evaluating expressions is slow

Expr should be avoided in time critical parts of applications. You will get a one time warning if you use them. They can be replaced easily by funs or function closures. They are evaluated at global scope in Module Main only. Other modules using DiscreteEvents.jl cannot use Expr in events and have to use functions.

source

DiscreteEvents.Timing — Type.

Timing

Enumeration type for scheduling events and timed conditions:

at: schedule an event at a given time,

after: schedule an event a given time after current time,

every: schedule an event every given time from now on,

before: a timed condition is true before a given time,

until: delay until t.

source

DiscreteEvents.fun — Function.

fun(f::Function, args...; kwargs...)

Save a function f and its arguments in a closure for later execution.

Arguments

The arguments args... and keyword arguments kwargs... to fun are passed to f at execution but may change their values between beeing captured in fun and fs later execution. If f needs their current values at execution time there are two possibilities:

fun can take funs, function closures, symbols or expressions at the place of values or variable arguments. They are evaluated at event time just before being passed to f. There is one exception: if f is an event!, its arguments are passed on unevaluated.

A mutable type argument (Array, struct ...) is always current. You can also change its content from within a function.

Evaluating symbols and expressions is slow

Symbols and Expr should be avoided in time critical parts of applications. You will get a one time warning if you use that feature. They can be replaced easily by funs or function closures. They are evaluated at global scope in Module Main only. Other modules using DiscreteEvents.jl cannot use this feature and have to use functions.

Examples

julia> using DiscreteEvents

julia> g(x; y=1) = x+y
g (generic function with 1 method)

julia> x = 1
1

julia> gg = fun(g, :x, y=2);   # we pass x as a symbol to fun

julia> x += 1   # a becomes 2
2

julia> gg()     # at execution g gets a current x and gives a warning
┌ Warning: Evaluating expressions is slow, use functions instead
└ @ DiscreteEvents ~/.julia/dev/DiscreteEvents/src/fclosure.jl:37
4

julia> hh = fun(g, fun(()->x), y=3);   # reference x with an anonymous fun

julia> x += 1   # x becomes 3
3

julia> hh()     # at execution g gets again a current x
6

julia> ii = fun(g, ()->x, y=4);  # reference x with an anonymous function

julia> x += 1   # x becomes 4
4

julia> ii()     # ok, g gets an updated x
8

source

DiscreteEvents.event! — Function.

event!([clk], ex, t; <keyword arguments>)
event!([clk], ex, T, t; <keyword arguments>)

Schedule an event for a given simulation time.

Arguments

clk<:AbstractClock: clock, it not supplied, the event is scheduled to 𝐶,

ex<:Action: an expression or function or a tuple of them,

T::Timing: a timing, one of at, after or every,

t<:Number: simulation time, if t < clk.time set t = clk.time,

Keyword arguments

cycle<:Number=0.0: repeat cycle time for an event,

cid::Int=clk.id: if cid ≠ clk.id, assign the event to the parallel clock with id == cid. This overrides spawn,

spawn::Bool=false: if true, spawn the event at other available threads,

sync::Bool=false: if true, force a synchronization of all parallel clocks before executing the event.

Examples

julia> using DiscreteEvents, Unitful

julia> import Unitful: s, minute, hr

julia> myfunc(a, b) = a+b
myfunc (generic function with 1 method)

julia> event!(𝐶, fun(myfunc, 1, 2), 1) # a 1st event to 1

julia> event!(𝐶, fun(myfunc, 2, 3), 1) #  a 2nd event to the same time

julia> event!(𝐶, fun(myfunc, 3, 4), 1s)
Warning: clock has no time unit, ignoring units

julia> setUnit!(𝐶, s)
0.0 s

julia> event!(𝐶, fun(myfunc, 4, 5), 1minute)

julia> event!(fun(myfunc, 5, 6), after, 1hr)

source

event!([clk], ex, cond; <keyword arguments>)

Schedule ex as a conditional event, conditions cond get evaluated at each clock tick.

Arguments

clk<:AbstractClock: if no clock is supplied, the event is scheduled to 𝐶,

ex<:Action: an expression or function or a tuple of them,

cond<:Action: a condition is true if all functions or expressions therein return true,

cid::Int=clk.id: assign the event to the parallel clock cid. This overrides spawn,

spawn::Bool=false: if true, spawn the event at other available threads.

Examples

julia> using DiscreteEvents

julia> c = Clock()   # create a new clock
Clock 0, thrd 1 (+ 0 ac): state=DiscreteEvents.Undefined(), t=0.0 , Δt=0.01 , prc:0
  scheduled ev:0, cev:0, sampl:0

julia> event!(c, fun((x)->println(tau(x), ": now I'm triggered"), c), fun(>=, fun(tau, c), 5))

julia> run!(c, 10)   # sampling is not exact, so it takes 502 sample steps to fire the event
5.009999999999938: now I'm triggered
"run! finished with 0 clock events, 502 sample steps, simulation time: 10.0"

source

Functions and expressions can be given to events on their own or in tuples, even mixed:

function events()
    event!(:(i += 1), after, 10)  # one expression
    event!(fun(f, 1, 2, 3, diff=pi), every, 1)  # one fun
    event!((:(i += 1), fun(g, j)), (()->tau() ≥ 50, fun(isready, input), :(a ≤ 10))) # two funs under three conditions
end

Events are called or evaluated at a their scheduled times or by sampling when their preconditions become true.

Note

For conditions you should prefer inequalities like <, ≤, ≥, > to equality == in order to make sure that a condition can be detected, e.g. tau() ≥ 100 is preferable to tau() == 100.

Continuous sampling

Functions or expressions can register for sampling and are then executed "continuously" at each clock increment Δt. The default clock sample rate Δt is 0.01 time units.

DiscreteEvents.sample_time! — Function.

sample_time!([clk::Clock], Δt::N) where {N<:Number}

Set the clock's sample rate starting from now (tau(clk)).

Arguments

clk::Clock: if not supplied, set the sample rate on 𝐶,

Δt::N: sample rate, time interval for sampling

source

DiscreteEvents.periodic! — Function.

periodic!([clk], ex, Δt; spawn)

Register a function or expression for periodic execution at the clock`s sample rate.

Arguments

clk<:AbstractClock: if not supplied, it registers on 𝐶,

ex<:Action: an expression or function or a tuple of them,

Δt<:Number=clk.Δt: set the clock's sampling rate, if no Δt is given, it takes the current sampling rate, if that is 0, it calculates one,

spawn::Bool=false: if true, spawn the periodic event to other available threads.

source

Processes

Julia functions can run as processes (asynchronous tasks) registered to a clock.

DiscreteEvents.Prc — Type.

Prc(id, f, arg...; kw...)

Prepare a function to run as a process (asynchronous task) in a simulation.

Arguments, fields

id: some unique identification for registration,

f::Function: a function f(clk, arg...; kw...), must take clk (a Clock) as its first argument,

arg...: further arguments to f

kw...: keyword arguments to f

Fields

task::Union{Task,Nothing}: a task structure,

clk::Union{AbstractClock,Nothing}: clock where the process is registered,

state::ClockState: process state,

Note

A function started as a Prc runs in a loop. It has to give back control by e.g. doing a take!(input) or by calling delay! or wait!, which will yield it. Otherwise it will starve everything else!

source

DiscreteEvents.process! — Function.

process!([clk], prc, cycles; spawn)

Register a Prc to a clock, start it as an asynchronous process and return the id it was registered with. It can then be found under clk.processes[id].

Arguments

c<:AbstractClock: if not provided, the process runs under 𝐶,

prc::Prc: it contains a function and its arguments,

cycles<:Number=Inf: number of cycles the process should run,

spawn::Bool=false: if true, the process may be scheduled on another thread in parallel and registered to the thread specific clock.

Note

spawning a process is possible only with parallel clocks setup with PClock or fork!.

source

DiscreteEvents.interrupt! — Function.

interrupt!(p::Prc, ev::ClockEvent, value=nothing)

Interrupt a Prc by throwing a ClockException to it.

source

Delay and wait …

Processes then can call delay! or wait! or take! and put! … on their channels. They are then suspended until a given time or until certain conditions are met or requested resources are available.

DiscreteEvents.delay! — Function.

delay!(clk, Δt)
delay!(clk, T, t)

Delay (suspend) a process for a time interval Δt on the clock clk.

Arguments

Δt<:Number: time interval,

T::Timing: only until is accepted,

t<:Number: delay until time t if t > clk.time, else give a warning.

source

DiscreteEvents.wait! — Function.

wait!(clk, cond)

Delay (suspend) a process on a clock clk until a condition has become true.

Arguments

cond<:Action: a condition is is true if all expressions or functions therein return true.

source

Now

Processes in a simulation enclose IO-operations in a now! call to make sure they finish before the clock proceeds.

DiscreteEvents.now! — Function.

now!(clk::Clock, ex::A) where {A<:Action}

Tell the clock to execute an IO-operation ex and not to proceed before ex is finished.

source

Running simulations

Virtual clocks can be run, stopped or stepped through and thereby used to simulate chains of events.

DiscreteEvents.run! — Function.

run!(clk::Clock, duration::N) where {N<:Number}

Run a simulation for a given duration.

source

DiscreteEvents.incr! — Function.

incr!(clk::Clock)

Take one simulation step, execute the next tick or event.

source

DiscreteEvents.resetClock! — Function.

resetClock!(clk::Clock, Δt::T=0.01; t0::U=0; <keyword arguments>) where {T<:Number, U<:Number}

Reset a clock.

Arguments

clk::Clock

Δt::T=0.01: sample rate

t0::Float64=0 or t0::Time: start time

hard::Bool=true: time is reset, all scheduled events and sampling are deleted. If hard=false, then only time is reset, event and sampling times are adjusted accordingly.

unit=NoUnits: the Time unit for the clock after reset. If a Δt::Time is given, its Time unit goes into the clock Time unit. If only t0::Time is given, its Time unit goes into the clock time unit.

Examples

julia> using DiscreteEvents, Unitful

julia> import Unitful: s

julia> c = Clock(1s, t0=60s)
Clock 0, thread 1 (+ 0 ac): state=DiscreteEvents.Undefined(), t=60.0 s, Δt=1.0 s, prc:0
  scheduled ev:0, cev:0, sampl:0

julia> resetClock!(c)
"clock reset to t₀=0.0, sampling rate Δt=0.01."

julia> c
Clock 0, thread 1 (+ 0 ac): state=DiscreteEvents.Idle(), t=0.0 , Δt=0.01 , prc:0
  scheduled ev:0, cev:0, sampl:0

source

resetClock!(rtc::RTClock)

Reset a real time clock. Set its time to zero and delete all scheduled and sampling events.

source

DiscreteEvents.stop! — Function.

stop!(clk::Clock)

Stop a running simulation.

source

Stop a Prc

source

DiscreteEvents.resume! — Function.

resume!(clk::Clock)

Resume a halted simulation.

source

DiscreteEvents.sync! — Function.

sync!(clk::Clock, to::Clock=𝐶)

Force a synchronization of two clocks. Change all registered times of clk accordingly. Convert or force clk.unit to to.unit.

source

Resources

Shared resources with limited capacity are often needed in simulations.

One approach to model them is to use Julia Channels with its API. This is threadsafe and thus should be preferred for multithreading applications.

Using Resource is a second possibility to model shared resources. Its interface gives more flexibility and is faster in single threaded applications, but in multithreading the user must avoid race conditions by explicitly wrapping access with lock -… access …- unlock.

DiscreteEvents.Resource — Type.

Resource{T}(capacity)

A Resource implements a Deque with a limited capacity. If used in multithreading applications, the user must avoid race conditions by explicitly wrapping modifying calls with lock-unlock.

Fields

items::Deque{T}: resource buffer

capacity::Number=Inf: the capacity is limited to the given integer,

lock::ReentrantLock: a lock for coordinating resource access by tasks.

Example

julia> 

Note

In order to use the full interface to Resource you have to load DataStructures.

source

DataStructures.capacity — Function.

capacity(r::Resource)

Get the capacity of a resource

source

capacity(ch::Channel)

Get the capacity of a channel.

source

DataStructures.isfull — Function.

isfull(r::Resource)

Test whether the resource is full

source

isfull(ch::Channel)

Test whether a channel is full.

source

Base.isready — Function.

isready(r::Resource)

Test whether an item is available.

source

Base.isempty — Function.

isempty(r::Resource)

Test whether the resource is empty.

source

Base.empty! — Function.

empty!(r::Resource)

Reset the resource buffer (deque).

source

empty!(ch::Channel)

Reset a channel, throw away the elements stored to it.

source

Base.length — Function.

length(r::Resource)

Get the number of elements available.

source

length(ch::Channel)

Get the number of items in a channel.

source

Base.push! — Function.

push!(r::Resource, x)

Add an element to the back of a resource deque.

source

Base.pop! — Function.

pop!(r::Resource)

Remove an element from the back of a resource deque.

source

Base.pushfirst! — Function.

pushfirst!(r::Resource, x)

Add an element to the front of a resource deque.

source

Base.popfirst! — Function.

popfirst!(r::Resource)

Remove an element from the front of a resource deque.

source

Base.first — Function.

first(r::Resource)

Get the element at the front of a resource deque.

source

first(ch::Channel)

Get the first item in a channel without removing it. Throws an error if the channel is empty. If you want a blocking operation, use fetch.

source

Base.last — Function.

last(r::Resource)

Get the element at the back of a resource deque.

source

Resource provides a lock-unlock API for multithreading applications.

Base.lock — Function.

lock(r::Resource)

Acquire the resource lock when it becomes available. If the lock is already locked by a different task/thread, wait for it to become available.

Each lock must be matched by an unlock.

source

Base.unlock — Function.

unlock(r::Resource)

Releases ownership of the resource lock.

If this is a recursive lock which has been acquired before, decrement an internal counter and return immediately.

source

Base.islocked — Function.

islocked(r::Resource)

Check whether the lock is held by any task/thread. This should not be used for synchronization (see instead trylock).

source

Base.trylock — Function.

trylock(r::Resource)

Acquire the resource lock if it is available, and return true if successful. If the lock is already locked by a different task/thread, return false.

Each successful trylock must be matched by an unlock.

source

Utilities

DiscreteEvents.onthread — Function.

onthread(f::F, tid::Int; wait::Bool=true) where {F<:Function}

Execute a function f on thread tid (to speed it up).

Arguments

f::Function: function to execute

tid::Int: thread id

wait::Bool=true: if true, it waits for function to finish

Example

julia> using DiscreteEvents, .Threads

julia> onthread(2) do; threadid(); end
2

source